Switch to long

+++ BetterRandom.hpp
@@ -15,12 +15,12 @@ extern "C"
 class BetterRandom
 {
 public:
-    BetterRandom(unsigned int);
+    BetterRandom(unsigned long);
     ~BetterRandom();
 
     static double get_rand_01(void); // Must return double in [0, 1)
-    static unsigned int get_rand_bits(void); // Must return unsigned int in [0, 2^32 - 1]
-    static unsigned int get_rand_tu01(void); // Get random number from TestU01 generator
+    static unsigned long get_rand_bits(void); // Must return unsigned int in [0, 2^32 - 1]
+    static unsigned long get_rand_tu01(void); // Get random number from TestU01 generator
 
     static void test_gen_01(void); // Perform TestU01 test battery on custom gen (double)
     static void test_gen_bits(void); // Perform TestU01 test battery on custom gen (uint)
@@ -35,7 +35,7 @@ private:
     static unif01_Gen* _gen_tu01; // TestU01 gen
     static bool _gens_deleted; // To avoid segfaulting by deleting twice
 
-    static unsigned int _last_num; // TODO Last generated value
+    static unsigned long _last_num; // TODO Last generated value
 
     static uint32_t _x, _y, _z, _w; // For Xorshift
