+++ CppRandom.hpp
@@ -11,12 +11,12 @@ extern "C"
 class CppRandom
 {
 public:
-    CppRandom(unsigned int);
+    CppRandom(unsigned long);
     ~CppRandom();
 
     static double get_rand_01(void);
-    static unsigned int get_rand_bits(int); // Get random uint from given gen index
-    static unsigned int get_rand_bits_wrapper(); // Wrapper to pass to TU01 test
+    static unsigned long get_rand_bits(int); // Get random uint from given gen index
+    static unsigned long get_rand_bits_wrapper(); // Wrapper to pass to TU01 test
 
     static void test_gen_01(void); // Perform TestU01 battery test on c++ gen (double)
     static void test_gen_bits(void); // Perform TestU01 battery test on c++ gen (uint)
